[feat] add old install file check

+++ src-tauri/src/integrator/ue4ss_integrate.rs
@@ -1,18 +1,28 @@
 use crate::integrator::ReadSeek;
+use std::error::Error;
 use std::fs;
 use std::path::PathBuf;
-use tracing::info;
 
 pub fn install_ue4ss(install_path: &PathBuf) {
     let ue4ss_path = install_path.join("ue4ss");
-    fs::create_dir(&ue4ss_path).unwrap();
+    println!("{:?}", ue4ss_path);
+
+    if !ue4ss_path.exists() {
+        fs::create_dir(&ue4ss_path).unwrap();
+    }
 
     let dll_path = ue4ss_path.join("UE4SSL.dll");
-    let ue4ss_dll = include_bytes!("../../assets/UE4SSL.dll");
-    fs::write(dll_path, ue4ss_dll).unwrap();
+    if !dll_path.exists() {
+        let ue4ss_dll = include_bytes!("../../assets/UE4SSL.dll");
+        fs::write(dll_path, ue4ss_dll).unwrap();
 
-    let mods_path = ue4ss_path.join("mods");
-    fs::create_dir(mods_path).unwrap();
+        let proxy_dll_path = install_path.join("dwmapi.dll");
+        let proxy_dll = include_bytes!("../../assets/dwmapi.dll");
+        fs::write(proxy_dll_path, proxy_dll).unwrap();
+
+        let mods_path = ue4ss_path.join("mods");
+        fs::create_dir(mods_path).unwrap();
+    }
 }
 
 pub fn install_ue4ss_mod(
@@ -32,25 +42,27 @@ pub fn install_ue4ss_mod(
     fs::write(dll_path, &content).unwrap();
 }
 
-pub fn uninstall_ue4ss(install_path: &PathBuf) {
+pub fn uninstall_ue4ss(install_path: &PathBuf) -> Result<(), Box<dyn Error>> {
     let ue4ss_path = install_path.join("ue4ss");
-    if fs::exists(&ue4ss_path) {
+    if fs::exists(&ue4ss_path)? {
         fs::remove_dir_all(ue4ss_path).unwrap();
     }
 
     //try to delete other ue4ss file
     let ue4ss_dll = install_path.join("UE4SS.dll");
-    if fs::exists(&ue4ss_dll) {
+    if fs::exists(&ue4ss_dll)? {
         fs::remove_file(&ue4ss_dll).unwrap();
     }
 
     let proxy_dll = install_path.join("dwmapi.dll");
-    if fs::exists(&proxy_dll) {
+    if fs::exists(&proxy_dll)? {
         fs::remove_file(&proxy_dll).unwrap();
     }
 
     let ue4ss_mods = install_path.join("mods");
-    if fs::exists(&ue4ss_mods) {
+    if fs::exists(&ue4ss_mods)? {
         fs::remove_dir_all(&ue4ss_mods).unwrap();
     }
+
+    Ok(())
 }
